refactor(ai): move the three /pages/ai routes into src/ai/page_routes.py
Pure move, no behaviour change. GET /pages/ai/context,
POST /pages/ai/generate and POST /pages/ai/chat were the last three
untagged, non-deprecated routes still inline in src/api_server.py.

Moved with them: the four helpers that were private to those handlers
and referenced by nothing else in the codebase — the generate throttle
(_AI_GENERATE_SEMAPHORE, _ai_generate_throttle_check), the SSE frame
writer (_format_sse_event), and the two prompt-context collectors
(_collect_ai_variables, _collect_plugin_demos). api_server drops 323
lines and gains an include_router call.

The new router is deliberately **untagged** in this commit. `ai` is
already in conventions_manifest.json's converted_domains, so the moment
these routes carry the `ai` tag the ratchet enforces all four rules on
them — which they do not satisfy yet. The tag arrives in the next
commit, together with the conventions, so no intermediate commit is red.

Seams: get_config_manager is imported from src.config_manager rather
than resolved through api_server, so the router imports nothing from the
app module. tests/test_api_ai_endpoints.py now stubs both targets — its
/settings/ai cases still go through the api_server seam, which cannot be
retired until the last consumer converts.

Verified:
- tests/golden/api_routes.json byte-identical (git diff --quiet: clean)
- full suite 6599 passed, 1 failed (test_check_image_formats
  TestRepositoryIsClean, the known worktree-environmental failure),
  2 skipped — a strict superset of the 6564 before this branch
- data-dir leak count 0
